- Se cambio precios a double en comprarion

+++ app/src/main/java/com/example/smartcompra/data/models/Producto.kt
@@ -4,11 +4,11 @@ data class Producto(
     val nombre: String,
     val marca: String?,
     val cantidad: Int,
-    val precio: Int,
+    val precio: Double,
     val unidad: String,
     val descuento: Int,
     val pack: Int,
-    val precioNormalizado: Int,
+    val precioNormalizado: Double,
     val unidadNormalizada: String,
     val bestPrice: Boolean = false
 )
